feat: add code editor to problem editor
